backend_congrats/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0 
pydantic_settings==2.10.1
weasyprint==66.0
